maindir/requirements.txt
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
pefile==2021.9.3
Pillow==8.4.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytils==0.3
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
MarkupSafe==2.1.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
django-jsonfield-backport==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
defusedxml==0.7.1
Django==3.2.9
django-cleanup==6.0.0
django-currentuser==0.5.3
django-debug-toolbar==3.2.4
altgraph==0.17.2
amqp==5.1.1
asgiref==3.4.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.6
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
